+++ src/main/kotlin/com/rafeng/bakery/improve/business/notification/filling/event/CreateFillingEvent.kt
@@ -1,4 +1,4 @@
-package com.rafeng.bakery.improve.business.notification.event
+package com.rafeng.bakery.improve.business.notification.filling.event
 
 import com.rafeng.bakery.improve.business.model.Filling
 import org.springframework.context.ApplicationEvent
