+++ src/main/kotlin/com/rafeng/bakery/improve/business/notification/filling/event/GetAllFillingsEvent.kt
@@ -1,6 +1,5 @@
-package com.rafeng.bakery.improve.business.notification.event
+package com.rafeng.bakery.improve.business.notification.filling.event
 
-import com.rafeng.bakery.improve.business.model.Filling
 import org.springframework.context.ApplicationEvent
 
 class GetAllFillingsEvent(listOfFillingIds: List<Long>, val message: String): ApplicationEvent(listOfFillingIds)
