+++ src/main/kotlin/com/rafeng/bakery/improve/business/notification/topping/event/CreateToppingEvent.kt
@@ -1,7 +1,6 @@
-package com.rafeng.bakery.improve.business.notification.event
+package com.rafeng.bakery.improve.business.notification.topping.event
 
 import com.rafeng.bakery.improve.business.model.Topping
-import com.rafeng.bakery.improve.business.model.entity.ToppingEntity
 import org.springframework.context.ApplicationEvent
 
 class CreateToppingEvent(toppingEntity: Topping, val message: String) : ApplicationEvent(toppingEntity)
