+++ src/main/kotlin/com/rafeng/bakery/improve/business/notification/topping/event/GetAllToppingsEvent.kt
@@ -1,6 +1,5 @@
-package com.rafeng.bakery.improve.business.notification.event
+package com.rafeng.bakery.improve.business.notification.topping.event
 
-import com.rafeng.bakery.improve.business.model.Topping
 import org.springframework.context.ApplicationEvent
 
 class GetAllToppingsEvent(listOfToppingIds: List<Long>, val message: String) : ApplicationEvent(listOfToppingIds)
