+++ src/main/kotlin/com/rafeng/bakery/improve/business/notification/topping/listener/DeleteToppingEmailEventListener.kt
@@ -1,6 +1,6 @@
-package com.rafeng.bakery.improve.business.notification.listener
+package com.rafeng.bakery.improve.business.notification.topping.listener
 
-import com.rafeng.bakery.improve.business.notification.event.DeleteToppingEvent
+import com.rafeng.bakery.improve.business.notification.topping.event.DeleteToppingEvent
 import org.springframework.context.ApplicationListener
 import org.springframework.stereotype.Component
